+++ CHANGELOG.md
@@ -5,11 +5,14 @@
 ### Added
 
 - All functions provided by the NewRelic PHP extension are now supported in the `NewRelicInteractorInterface`.
-- Added a new `deprecations` parameter to logs `E_USER_DEPRECATED`. 
+- Added a new `deprecations` parameter to logs `E_USER_DEPRECATED`.
+- Added a new `monolog` parameter to send logs to new relic.
 
 ### Changed
 
 - Command Configuration explicit
+- The configuration syntax
+- The bundle uses class-named service ids. See `UPGRADE-2.0.md` for the exhaustive list of changes
 
 ### Removed